renamed order_manager hpp and cpp files
