+++ vda5050_core/include/vda5050_core/order_execution/order.hpp
@@ -23,8 +23,8 @@
 #include <string>
 #include <vector>
 
-#include "vda5050_core/order_execution/Edge.hpp"
-#include "vda5050_core/order_execution/Node.hpp"
+#include "vda5050_core/order_execution/edge.hpp"
+#include "vda5050_core/order_execution/node.hpp"
 
 namespace vda5050_core {
