+++ vda5050_core/src/vda5050_core/order_execution/edge.cpp
@@ -19,8 +19,8 @@
 #include <cstdint>
 #include <string>
 
-#include "vda5050_core/order_execution/Edge.hpp"
-#include "vda5050_core/order_execution/OrderGraphElement.hpp"
+#include "vda5050_core/order_execution/edge.hpp"
+#include "vda5050_core/order_execution/order_graph_element.hpp"
 
 namespace vda5050_core {
 namespace edge {
