+++ vda5050_core/src/vda5050_core/order_execution/node.cpp
@@ -18,8 +18,8 @@
 
 #include <string>
 
-#include "vda5050_core/order_execution/Node.hpp"
-#include "vda5050_core/order_execution/OrderGraphElement.hpp"
+#include "vda5050_core/order_execution/node.hpp"
+#include "vda5050_core/order_execution/order_graph_element.hpp"
 
 namespace vda5050_core {
