+++ vda5050_core/src/vda5050_core/order_execution/order_graph_element.cpp
@@ -19,7 +19,7 @@
 #include <cstdint>
 #include <string>
 
-#include "vda5050_core/order_execution/OrderGraphElement.hpp"
+#include "vda5050_core/order_execution/order_graph_element.hpp"
 
 namespace vda5050_core {
 namespace order_graph_element {
