+++ vda5050_core/src/vda5050_core/order_execution/order_graph_validator.cpp
@@ -20,9 +20,9 @@
 #include <queue>
 #include <vector>
 
-#include "vda5050_core/order_execution/Edge.hpp"
-#include "vda5050_core/order_execution/Node.hpp"
-#include "vda5050_core/order_execution/OrderGraphValidator.hpp"
+#include "vda5050_core/order_execution/edge.hpp"
+#include "vda5050_core/order_execution/node.hpp"
+#include "vda5050_core/order_execution/order_graph_validator.hpp"
 
 namespace vda5050_core {
 namespace order_graph_validator {
